chore: merged downloader with UI (function and button not binded yet)
